tests/wpt/tests/css/css-view-transitions/no-crash-set-exception.html

<!DOCTYPE html>
<html>
<title>View transitions: author styles ignored during prepare</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-1/">
<link rel="author" href="mailto:khushalsagar@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  contain: paint;
  view-transition-name: shared;
}

html::view-transition,
html::view-transition-group(shared),
html::view-transition-image-pair(shared),
html::view-transition-old(shared),
html::view-transition-new(shared) {
  background: blue;
}
</style>

<div></div>

<script>
promise_test(async t => {
  assert_implements(document.startViewTransition, "Missing document.startViewTransition");
  return new Promise((resolve, reject) => {
    document.startViewTransition(() => {
      resolve();
      throw 'error';
    });
  });
}, "An exception thrown during a transition shouldn't crash.");

</script>
